cyclictest: use --measurement-cpulist to place cyclictest threads
Change to use the selected set of cpu cores for measuremen threads
rather than assuming all cores will be measured. This means that
the cpu list may be 'sparse' so take care to not use ranges; use
the actual list of selected cpu cores.

Stop parsing /proc/cpuinfo and use the helper routine cpuinfo() to
get cpu information.

Add a text representation for a RunData object so that it
can be printed (mostly for debugging).

Change the report generation loop to iterate over the actual cpus
specified and not a range of 0..ncpus since a range no longer
specifies the correct cpus.

Signed-off-by: Clark Williams <williams@redhat.com>