kcompile.py: modify to manage jobs on per-node basis
This modification to kcompile takes into account NUMA topology
rather than treating the system as a big SMP box. Make loads are
started on a per-node basis (e.g. on a four node system there would be
four makes started), and the make commands are restricted to the cpu
cores available on that node.

This is an attempt to reduce cross-node traffic and to better mimic
a well-behaved realtime application.

Signed-off-by: Clark Williams <williams@redhat.com>