loads/hackbench: Fix number of jobs calculation
Number of hackbench jobs are calculated w.r.t. the number of CPUs the
load is allowed to run on. When restricting such CPU mask by specifying
a cpulist and int to str check is however wrongly performed and this
causes the calculation to return 0 (making hackbench failing to start).

Fix it by performing a conversion to string before comparing against the
cpulist.

Signed-off-by: Juri Lelli <juri.lelli@redhat.com>
Signed-off-by: John Kacur <jkacur@redhat.com>