cyclictest: fix start options to correctly invoke thread/affinity options
Behavior change in cyclictest command line (removal of --numa/--smp) caused
issue when running where only one cpu thread was started when rteval thought
that a thread per online cpu was started. This caused a crash when the run
was finished and the histogram data was being processed. Change the logic
to not use --numa/--smp and when no cpulists specified just use '-t -a' to
put a measurement thread on each cpu.

Signed-off-by: Clark Williams <williams@redhat.com>

Note, the patch seems to be correct, but --smp was not removed from
cyclictest, just --numa was removed
Signed-off-by: John Kacur <jkacur@redhat.com>