rteval: Rename self.cmd to self.args for inheritance in stressng
Load modules inherits from CommandLineLoad
CommandLineLoad has a method MakeReport which relies on the variable
"args". Class Stressng needs to use this variable name instead of it's
own for this mechanism to work.

This commit changes the name in Stressng, but perhaps a better fix in
the future would be to use accessor methods, instead of relying on a
variable name

This commit also sets jobs to 1, which is required in MakeReport. The
actual number of jobs is handled by the external stress-ng program,
depending upon the arg that Class Stressng invokes it with.

Signed-off-by: John Kacur <jkacur@redhat.com>