rteval:unittest.py Apply some fixes to unittest.py
The unittest system in rteval is old, and broken, and not extensive
enough. It should be removed and replaced with the the unittest
facilities in python3 and expanded. However, rather than just throwing
away what we have, I would like to take a look at what is there, and see
what can be used and imported into a new scheme. With that in
mind, here is a partial fix to the unittest.py that improves the
results.

Signed-off-by: John Kacur <jkacur@redhat.com>