rteval: Remove requirement to install python-schedutils
This is actually a no-op in the code as no code was actually importing
python-schedutils or using it!

This patches removes mention of it from a few documentation files.

Note: we should remove rteval.spec entirely, the reason we cannot yet
is that there is some code that parses the version number from it.

Signed-off-by: John Kacur <jkacur@redhat.com>